Initializer and Volunteer Shift change
Initializer
Increased the amount of events to 800.  More future events.
Volunteer now have a shift on one day instead of being able to have several from the initializer.  Not validation.

Volunteer master details
Changed volunteer details to show case shifts from now to the next 7 days.  No need for paging since the max is 7 shifts due to validation and the setup of the initialized data